Stubs changed into examples. Java class moved.
