+++ examples/jni-lib/CMakeLists.txt
